🐛 Bug fixes (Youtube / Music bot player)
- Removed message deletion
This was causing a semi-annoying 'missing message' issue since we deleted the message once we used play
- Fixed annoying NULL entry on db
Since we can't remove an individual index by popping it, the weird unsert pull thing has to stay but that can leave a NULL entry. We just check for the NULL on the top level and in arrays
- Added table,query,queue into catch block and dispatcher listener (so they're updated and not pulling from the functions variables)
- Allowing to play videos if the server has a queue in DB (Quality of life change more than anything)

- Removed networking interfacing on index.js
Turns out heroku dynamically changes ip on every redeployment so grapping IP off of what we were logging doesn't even really do anything.