Quality of life updates
`delete
Replaced
- Added alias 'clean'
Reworked delete to not use bulk delete and to delete in a loop (yes it's slow) but was only real way to get around the '14' day restriction on bulk deletion.

`changes
- Added time-ago
Added  time ago to the changes to see how long ago the bot was updated (has date as well but is easier to read at a glance)

`commands
Updated commands to reflect changes to delete

`monkey
- Moved voting system into a try catch.
- Removed

`invite
- Added (click me) so more people know to click the link

`uptime
- Replace all the code with time ago

Error
- Updated error to tell me when error occured (time-ago)

Command Stuff
- Updated rate liting to only activate on commands
Was targeting on prefix so database was filling up for no good reason.