👷 Error 429 and you.
Error 429 is a HTTP error response for rate limiting. We're getting it. ytdl-core is currently returning 429 errors due to high traffic and I don't believe there's a solid solution to getting around this currently. So for the time being I am returning a message saying that we reached the limit to any user that encounters it.

It also looks like Node cannot handle SIGTERM singles correctly. I wished to implement a reader for SIGTERM terminations to exit some what safely but it appears my efforts are thwarted.

- Attempted to catch error 429 from ytdl-core mini-get
- Updated error handler to no longer message the chat.
- Updated error handler to exit after 5 seconds (we need discord to send the message)
- Fixed time ago typo
- Fixed same voice channel bug
- Fixed no thumbnail for category gaming
- Added debugging status to bot
- Added automatic disabling of rate limiting while in debug mode
- Started development on youtube searching
-