Future of the bot in the next coming weeks
Future of the bot is going to be slow in the next coming weeks. This is due to UNI wrapping up for this semester and big plans I have for the bot going forward. My main priority right now is fixing what was previously fine and optimal code to be more efficient as a noticeable increase in computing time has become apparent which leads me to my second plan. Upgrading monkey bot to run faster. Now the only real problem with this is I am not rich by any means and cannot pocket out any money at this time so I need to plan a finical way that benefits both user and pays for monkey bot in a way that doesn’t completely ruin the functionality of the bot for normal people everyday users. (I do not want monkey bot to become a pay to use thing) So I have come up with ‘If they pay for the extra computing power, they should probably get it’ approach. So, the road map for monkey bot is basically:
-	Fix the obvious bugs and inefficiencies
-	Storing monkey images / ranks (We can then search by rank)
-	The big one: Adding website functionality
-	Maybe investigate brain.js for monkey image identification? (Complete overkill but funny)

-- This update
• Added `suggest (since people where reporting suggestions)
• Added a `reply feature for the developer (Can reply to suggestions)
• Minor DB changes
• Minor quality of life changes