⚡️ Code optimization and Memory mapping
Exams are over, time to work on the bot.
I added heapdump library (was using nodejs V8 dump but heard good things about this library). This allows me to view the memory usage of the code. We are currently using a lot on arrays (not surprising), but this is not due to my code but dependencies we are running. On top of this I have cleaned up some code that was semi-redundant due to .flat() existing.

My plan for the coming days.
- Add worker threads .
- Implement the ability to add search terms for monkey bot (with review).
- Figure out if there is a way to fix `play functionality (The bot is so big that we are constantly getting 429 (Too many requests) errors).
- Start work on website functionality.