🚧 Web stuff + bug fixes
Sorry for the late update on the bot. Was getting some weird errors with console logging throwing errors. Anyhow, bot just got a polish up a bit in this update since it’s mainly just been bug fixes due to a lack of time to actually just throw at this project.
Todo:
• Decide on a front-end renderer (Website)
• Develop a WebSocket for live logging (Website)
• Develop a RESTAPI for the website
Changes
• Fixed leader board bug
• Fixed queue bug
• Reallocated logging to separate file + timestamping
• Created a logger for the website