Small update
+ New command `count / `random
+ Added Monke Subreddit to search terms
+ Updated index to not ignore rate limiting in development environments
+ Added functionality to ban feature (I am  now able to revoke access to users from using monkey bot / report spamming)
+ Updated report to include user ID

Why the ban feature? Over the past couple months users would report issues, at first it was helpful stuff like trying to setup the bot on their server etc. Now a days I am just getting report spammed about monkey bot sending pictures that are not monkeys. I can 'fix' this by adding a neural network that can identify a picture of a monkey but that is too much work and out of scope for this hobby project.

So, the ban feature was added. Repeat offenders or people that have been noted to spam the report feature in quick succession will be permanently blacklisted from monkey bot.

In the future if the bot ends up sending 1 specific image at a time and a user reports it, I will make it a feature where monkey bot will only send that image to that user for the foreseeable future.

(Edit: Fixed unban, was using mention system)