🎥 Video support
Added video support to the bot! By default video playback is turned off. This just makes life easier on the DB end since we won't need to update all the servers to fit the new schemea and when people opt into the video feature their server will update to the new one.

So why the videos? It's been bugging me for a while that the bot was completely ignoring video content due to how discord embeds work. In a short sense you cannot put a video inside a Message embed. So the current implementation is a compromise between having the image of the video but having to exit the app to play it and been able to play the video in discord but not having a fancy embed. Another thing to add would be that videos are such a rare encounter for the bot that it'll take a while to debug this new feature. I've tested it in a dev environment where I was getting nothing but videos and am happy with the current iteration.

+ Added video support to monkey bot
+ Added video command to opt in / out of video features
+ Updated database schemes to reflect new changes