Minor Tweaks and Updating
> Removed Web activity / support
> Removed Youtube functionality
> Removed redundant packages
> Updated all packages
> Modified all sorting functions

Updated all sorting functionality to work with mapping instead this _should_ make the bot slightly faster when handling concurrent sorting across different servers. This is a minor tweak before a major re-write of the entire bot so don't expect to much in terms of quality of life / quality of bot in the coming weeks.

-- Forgot to remove some console logging