Major update and entire code base re-write - Updated all commands to slash commands - Massive performance increase -  Logic reworks TODO (re-add voting with buttons / select menu?)
After a year of no updates I finally had some time to work on this bot again. I've re-wrote all the code for the bot to now be in typescript. Should hopefully result in some nice performance upgrades and helps with reduncy checks that existed in JavaScript. I'm expecting the bot to fail on some servers that didn't allow admin since slash commands did not exist 1-2 years ago when the bot was most active. Anyways hopefully the stats talk for themselves on how massive of an update this was.

- Update for deployment  & formatting (1294 deletions previously and new files obviously)